pyproject.toml
[project]
name = "movie-request-service"
version = "0.1.0"
description = """
minimal movie request service where users can request movies which will be downloaded and stored on the server
"""
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "flask[async]",
    "flask_limiter",
    "httpx",
    "python-dotenv",
    "dataclasses_json",
    "libtorrent",
    "requests_toolbelt",
    "guessit",
]

[dependency-groups]
dev = ["pytest", "pytest-flask", "pytest-asyncio"]
prod = ["gunicorn"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = 'auto'
markers = [
    "require_qbittorrent: mark test as requiring qbittorrent to be running",
]
addopts = "-m 'not require_qbittorrent' --log-cli-level=INFO -s"
